added a controller for users
